requirements.txt
crewai
streamlit
pandas
openai
python-dotenv
fastapi
uvicorn
python-multipart 